[#D3D-4677] Intents to fix SSL issues for version check
On OSX and Linux, Blender is shipped with an old version of SSL
python lib. This version uses an old and deprecated TLS protocol
that is blocked by major websites/servers for security reasons.

Github is blocking this protocol so the plugin is not able to
check for new versions on OSX/Linux.
The fix is to override the SSL lib with a newer version when importing
modules in the plugin.